refactor: move geometric layers off views
Move all contents of `createXXXLayer` in XXXView to XXXLayer in
Prefab/XXX/. This declutters the XXXView files. This change applies for
Globe, Planar and Panorama.

In the same time, GeometryLayer has been extract from Layer and given
some new methods which were repeated through each View.
