+++ src/Core/Layer/Layer.js
@@ -1,5 +1,4 @@
 import { EventDispatcher } from 'three';
-import Picking from '../Picking';
 
 /**
  * Fires when layer sequence change (meaning when the order of the layer changes in the view)
@@ -52,51 +51,6 @@ export const defineLayerProperty = function defineLayerProperty(layer, propertyN
                     }
                 } });
     }
-};
-
-function GeometryLayer(id, object3d) {
-    if (!id) {
-        throw new Error('Missing id parameter (GeometryLayer must have a unique id defined)');
-    }
-    if (!object3d || !object3d.isObject3D) {
-        throw new Error('Missing/Invalid object3d parameter (must be a three.js Object3D instance)');
-    }
-    this._attachedLayers = [];
-
-    if (object3d && object3d.type === 'Group' && object3d.name === '') {
-        object3d.name = id;
-    }
-
-    Object.defineProperty(this, 'object3d', {
-        value: object3d,
-        writable: false,
-    });
-
-    Object.defineProperty(this, 'id', {
-        value: id,
-        writable: false,
-    });
-
-    // Setup default picking method
-    this.pickObjectsAt = (view, mouse, radius) => Picking.pickObjectsAt(view, mouse, radius, this.object3d);
-
-    this.postUpdate = () => {};
-}
-
-GeometryLayer.prototype = Object.create(EventDispatcher.prototype);
-GeometryLayer.prototype.constructor = GeometryLayer;
-
-GeometryLayer.prototype.attach = function attach(layer) {
-    if (!layer.update) {
-        throw new Error(`Missing 'update' function -> can't attach layer ${layer.id}`);
-    }
-    this._attachedLayers.push(layer);
-};
-
-GeometryLayer.prototype.detach = function detach(layer) {
-    const count = this._attachedLayers.length;
-    this._attachedLayers = this._attachedLayers.filter(attached => attached.id != layer.id);
-    return this._attachedLayers.length < count;
 };
 
 /**
@@ -179,4 +133,4 @@ const ImageryLayers = {
     },
 };
 
-export { GeometryLayer, Layer, ImageryLayers };
+export { Layer, ImageryLayers };
