+++ src/Core/MainLoop.js
@@ -1,5 +1,6 @@
 import { EventDispatcher } from 'three';
-import { GeometryLayer, Layer } from './Layer/Layer';
+import { Layer } from './Layer/Layer';
+import GeometryLayer from './Layer/GeometryLayer';
 import Cache from '../Core/Scheduler/Cache';
 
 export const RENDERING_PAUSED = 0;
