+++ src/Core/View.js
@@ -4,7 +4,8 @@ import Camera from '../Renderer/Camera';
 import MainLoop, { MAIN_LOOP_EVENTS, RENDERING_PAUSED } from './MainLoop';
 import c3DEngine from '../Renderer/c3DEngine';
 import { STRATEGY_MIN_NETWORK_TRAFFIC } from './Layer/LayerUpdateStrategy';
-import { GeometryLayer, Layer, defineLayerProperty } from './Layer/Layer';
+import { Layer, defineLayerProperty } from './Layer/Layer';
+import GeometryLayer from './Layer/GeometryLayer';
 import Scheduler from './Scheduler/Scheduler';
 import Picking from './Picking';
 import { updateLayeredMaterialNodeImagery, updateLayeredMaterialNodeElevation } from '../Process/LayeredMaterialNodeProcessing';
